pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "home-lab-topology-tools"
version = "4.0.0"
description = "Infrastructure-as-Data topology generator for Proxmox home lab"
readme = "README.md"
requires-python = ">=3.8,<4"
license = "MIT"
authors = [
    {name = "Dmitri"}
]

dependencies = [
    "pyyaml>=6.0",
    "jinja2>=3.1.0",
    "jsonschema>=4.20.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=23.9.1",
    "isort>=5.13.0",
    "pylint>=3.0.0",
    "mypy>=1.7.0",
    "types-pyyaml>=6.0",
    "types-jinja2>=2.11",
    "yamllint>=1.34.0",
    "pre-commit>=3.3.0",
    "detect-secrets>=1.4.0",
]

[tool.black]
line-length = 120

[tool.isort]
profile = "black"
line_length = 120

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["v4/tests"]
python_files = ["test_*.py", "*_test.py"]
addopts = "-v --cov=v4/topology-tools --cov-report=term-missing"

[tool.setuptools.packages.find]
where = ["v4/topology-tools"]
include = ["scripts", "scripts.*"]
